+++ test/Armada.Test.Unit/Program.cs
@@ -259,6 +259,7 @@ namespace Armada.Test.Unit
             runner.AddSuite(new ConsumerVesselResolverTests());
             runner.AddSuite(new VoyageCheckArmingPlanTests());
             runner.AddSuite(new CheckRunSummaryViewTests());
+            runner.AddSuite(new RescueEffectivenessTests());
             runner.AddSuite(new DiskLifecycleTests());
             runner.AddSuite(new CheckRunParsingServiceTests());
             runner.AddSuite(new MergeRefCandidateTests());